==> analysis/02_PreProcessing.Rmd <==
---
title: "Phyloseq Preprocessing"
author: "Gurpreet Kaur"
date: "`r Sys.Date()`"
output:
  html_document: 
    code_folding: show
    theme: spacelab
    highlight: pygments
    keep_md: no
    toc: yes
    toc_float:
      collapsed: no
      smooth_scroll: yes
      toc_depth: 3
  keep_md: true  
editor_options: 
  chunk_output_type: console
---

```{setup}
knitr::opts_chunk$set(echo = TRUE,
                      fig.align = "center",
                      fig.path = "../figures/02_PreProcessing/") # send any figure output to this folder 
```

# Load libraries 
```{r load-libraries}
library(devtools)
#devtools::install_github("joey711/phyloseq")
library(phyloseq)
library(tidyverse)
```

# Goals

Here, I am working on processing the data into a phyloseq object. 

- ASV table 
- Taxonomy Table 
- Track Reads (metadata)

Then, I will remove the following: 

1. Remove Chloroplasts
2. Remove Mitochondria. 
3. Remove samples without "enough" reads. 

Finally, write data file of phyloseq output.

# Load Data 

## ASV Table
```{r load-asv-table}
# First, load asv table
load("data/01_DADA2/ASV_counts.RData")

# Inspect asv_tab
head(asv_tab)[,1:5]

# Fix names 
sample_names <- colnames(asv_tab)
samples_fixed <- sapply(strsplit(basename(sample_names), "_"), `[`,1) 
head(samples_fixed)

# re-write the ASV count file to fix names 
colnames(asv_tab) <- samples_fixed
str(asv_tab)
```

##  Taxonomy Table
```{r load-tax-table}
tax_df <- read.table("data/01_DADA2/ASV_taxonomy.tsv", sep = "\t", skip = 1)
head(tax_df)

# fix column names 
colnames(tax_df) <- c("asv_names", "Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species", "ASV", "ASVseq")

head(tax_df)

# Taxonomy Table Matrix
tax_mat <- 
  tax_df %>%
  tibble::column_to_rownames(., var = "asv_names") %>%
  as.matrix()
```


##  Track Reads Data
```{r load-track-reads}
load("data/01_DADA2/track_read_counts.RData")

# Take a look at the data
head(track_counts_df)
dim(track_counts_df)

# Load in metadata
metadata_df_old <- read.csv("data/metadata.csv")
dim(metadata_df_old)

metadata_df <- data.frame(sample_type = c(metadata_df_old$sample_type))
metadata_df <- separate(metadata_df_old, sample_type, into = c("treatment", "year"), sep = "-")
metadata_df
metadata_df$year = factor(metadata_df$year, levels = c('1', '5', '10'))
colnames(metadata_df)
dim(metadata_df)


# Merge metadata_df with track_reads_df 
head(metadata_df)
head(track_counts_df)

metadata_track_reads_df <- 
  metadata_df %>%
  left_join(., track_counts_df, by = c("Run" = "names")) %>%
  rename(names = Run)


# Intuition check 
head(metadata_track_reads_df)

# Update row.names to be sample names 
## Before 
row.names(metadata_track_reads_df)
# Rewrite 
row.names(metadata_track_reads_df) <- metadata_track_reads_df$names
# Check afterwards that it worked 
row.names(metadata_track_reads_df)
# intuition check
head(metadata_track_reads_df)
```


# Handoff to phyloseq
```{r phyloseq-handoff}
# double check it's all good 
dim(asv_tab)
dim(tax_mat)

# Intuition check 
stopifnot(row.names(asv_tab) == row.names(tax_mat))

# Construct the phyloseq object 
raw_physeq <- phyloseq(otu_table(asv_tab, taxa_are_rows = TRUE),
                       sample_data(metadata_track_reads_df),
                       tax_table(tax_mat))
raw_physeq

# Save this raw phyloseq object 
save(raw_physeq, file = "data/02_PreProcessing/raw_physeq.RData")
```

# Clean up the data

Remove: 

1. Chloroplasts
2. mitochondria  

```{r rm-mitos-chloros}
library(dplyr)

# Make new physeq without chloroplasts
noChloros_physeq <- raw_physeq %>%
  # rm chloroplasts
  subset_taxa(Order != "Chloroplast" | is.na(Order))

# How many taxa were chloroplasts?
num_chloro_ASVs <- ntaxa(raw_physeq) - ntaxa(noChloros_physeq)
num_chloro_ASVs

# Intuition check
tax_table(noChloros_physeq) %>%
  data.frame() 

# Remove mitochondria
noChlorosMitos_physeq <- noChloros_physeq %>%
  subset_taxa(Family != "Mitochondria" | is.na(Family))

# How many mitochondrial ASVs?
num_mito_ASVs <- ntaxa(noChloros_physeq) - ntaxa(noChlorosMitos_physeq)
num_mito_ASVs

# How many total ASVs were removed from chloros and mitos?
ntaxa(raw_physeq) - ntaxa(noChlorosMitos_physeq)

# Proportion of ASVs kept?
ntaxa(noChlorosMitos_physeq) / ntaxa(raw_physeq)

```

# Evaulate and remove the control samples 
My project data does not have control samples and mock community, so I am not following any of these steps to remove them. I am moving directly to evaluate the sequencing depth.


# Evaluate the Sequencing Depth 

```{r seq-depth}
# The current data object
noChlorosMitos_physeq

# What is the library size/sequencing depth for each sample? 
seqSums_df <- 
  noChlorosMitos_physeq %>%
  otu_table() %>%
  # Sum each sample column 
  colSums() %>%
  data.frame() %>%
  rownames_to_column(var = "names") %>%
  left_join(., metadata_track_reads_df, by = "names") 

# Rename second column 
colnames(seqSums_df)[2] <- "TotalSeqs"

# check
dim(seqSums_df)
head(seqSums_df)

# Show the depth of samples 
seqSums_df %>%
  dplyr::select(names, TotalSeqs) %>%
  arrange(TotalSeqs) %>%
  head()

# plot it! 
seqSums_df %>%
  ggplot(aes(x=reorder(names, TotalSeqs), y = TotalSeqs,
             fill = treatment)) + 
  geom_bar(stat = "identity") 

# Density plot 
seqSums_df %>%
  ggplot(aes(TotalSeqs, fill = treatment)) +
  geom_density(alpha = 0.5)
```

# Remove samples with few reads 

```{r rm-samps}
# What's the min seq depth? 
min(sample_sums(noChlorosMitos_physeq))

# Remove SRR11364374
raw_preprocessed_physeq <- 
  noChlorosMitos_physeq %>%
  subset_samples(., names != "SRR11364374")

#What's the new min seq depth?
min(sample_sums(raw_preprocessed_physeq))
```

# Save Preprocessed Phyloseq Object
```{r save-physeq}
save(raw_preprocessed_physeq,
     file = "data/02_PreProcessing/raw_preprocessed_physeq.RData")
```


#Conclusion of this file:
In this file I learnt about making a phyloseq object and processing our data into it. Also, I learnt that we have to remove chloroplasts and mitochondria from our data. My data set did not have information about mock communities and negative control samples. It is also important to remove samples without enough reads.

# Session Information 
```{r session-info}
# Ensure reproducibility 
devtools::session_info()
```
