==> analysis/03b_Phylogenetic_Tree.Rmd <==
---
title: "Phylogenetic Tree Inspection and Rooting"
author: "Gurpreet Kaur"
date: "`r Sys.Date()`"
output:
  html_document: 
    code_folding: show
    theme: spacelab
    highlight: pygments
    keep_md: no
    toc: yes
    toc_float:
      collapsed: no
      smooth_scroll: yes
      toc_depth: 3
  keep_md: true  
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      fig.align = "center",
                      # Always relevant to the document directory 
                      # send any figure output to this folder 
                      fig.path = "../figures/03_Phylogenetic_Tree/",
                      warning = FALSE) 
```

# Goals 

1. Load the fastree unrooted tree.  
2. Add tree to phyloseq object.  
3. Visualize and inspect tree with ggtree. 
4. Prune ASVs, if needed.  
5. Root our tree. 
6. Combine new tree with a phyloseq object. 
7. Save 2 phyloseq objects: 1. Unrooted tree phyloseq object, 2. Rooted tree phyloseq object. 

# Before you start

## Set my seed 
```{r set-seed}
# Any number can be chosen 
set.seed(238427)
```


## Load Packages 
```{r load-packages}
pacman::p_load(tidyverse, phyloseq, ggtree, phytools,
               install = FALSE)
```

## Load Data files 
```{r load-data}
# Preprocessed phyloseq object 
load("data/02_PreProcessing/raw_preprocessed_physeq.RData")
raw_preprocessed_physeq

# Load in the tree! 
unrooted_tree <- read.tree("data/03_Phylogenetic_Tree/ASVs_unrooted.tree")
unrooted_tree
str(unrooted_tree)
```

# Merge Phyloseq 
```{r merge-physeq}
# Intuition check 
stopifnot(ntaxa(raw_preprocessed_physeq) == ntaxa(unrooted_tree))

# Merge the tree with the phyloseq object 
unrooted_physeq <- 
  merge_phyloseq(raw_preprocessed_physeq, unrooted_tree)
unrooted_physeq
```


# Plot Tree with `ggtree`

```{r plot-tree-unrooted}
# Make a basic tree
kingdom_tree <- 
  ggtree(unrooted_physeq) + 
  # color tips by kingdom 
  geom_tippoint(mapping = aes(color = Kingdom)) + 
  scale_color_manual(values = c("goldenrod1", "cornflowerblue", "grey")) +
  # Add title 
  labs(title = "Unrooted Tree") + 
  #move the legend to the bottom 
  theme(legend.position = "bottom"); kingdom_tree

kingdom_node_tree <- 
  kingdom_tree + 
  # Add the node label 
  geom_text(aes(label=node), hjust= -0.5, vjust = -0.3, size = 2)
kingdom_node_tree
```

# Midroot Tree
```{r midroot-tree}
# Is the tree rooted?
is.rooted(unrooted_tree)

# Let's midpoint root the tree
midpoint_rooted_tree <- midpoint.root(unrooted_tree)

# Is the new tree rooted?
is.rooted(midpoint_rooted_tree)

# Merge tree with the new physeq
midroot_physeq <- 
  merge_phyloseq(raw_preprocessed_physeq, midpoint_rooted_tree)
midroot_physeq


# Quick inspection of tree 
ggtree(midroot_physeq) + 
  geom_tippoint(mapping = aes(color = Kingdom))

```


# Save to a new phyloseq object
```{r save-physeq}
# Save both phyloseq objects with our tree object to one .RData file 
save(list = c("unrooted_physeq", "midroot_physeq"),
     file = "data/03_Phylogenetic_Tree/phytree_preprocessed_physeq.RData")
```
 
#Conclusion:
The unrooted tree from my dataset did not show any long branches. But it was good learn how proceed with both the cases if I encounter them in future.

# Session Information 
```{r session-info}
# Ensure reproducibility 
devtools::session_info()
```
