==> analysis/04_Biodiversity.Rmd <==
---
title: "Biodiversity Analysis"
author: "Gurpreet Kaur"
date: "`r Sys.Date()`"
output:
  html_document: 
    code_folding: show
    theme: spacelab
    highlight: pygments
    keep_md: no
    toc: yes
    toc_float:
      collapsed: no
      smooth_scroll: yes
      toc_depth: 3
  keep_md: true  
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      fig.path = "../figures/04_Biodiversity/",
                      fig.align = "center")
```

# Setting the Environment 

## Set the seed 
```{r set-seed}
set.seed(238427)
```

## Load Libraries 
```{r load-packages}
pacman::p_load(tidyverse, devtools, patchwork, iNEXT, phyloseq,
               install = FALSE)
```


## Load in Data 
```{r load-data}

load("data/02_PreProcessing/raw_preprocessed_physeq.RData")
raw_preprocessed_physeq

# Intiution Check 
min(sample_sums(raw_preprocessed_physeq))

# Make metadata dataframe
metadata_df <-
  raw_preprocessed_physeq %>%
  sample_data() %>%
  data.frame()

head(metadata_df)

# Setting colors for treatments to match with dataset from the paper
treatment_colors <- c(
  "N0" = "brown3",
  "N120" = "dodgerblue4",
  "N180" = "darkolivegreen4",
  "N240" = "gold",
  "N360" = "deeppink4")

```

# Goals

1. Calculate the Hill Diversity of the samples. 
2. Evaluate the rarefaction curves. 
3. Evaluate the Diversity values. 
4. Makes notes of specific samples and their seq depth. 

# Diversity Calculations with iNEXT 

```{r calc-div}
# prepare input data 
iNEXT_input_df <- 
  raw_preprocessed_physeq %>%
  otu_table() %>%
  data.frame()

# Quick check

dim(iNEXT_input_df)

# Run iNEXT: Calculate the Hill Numbers 
# Note that: Species in ROWS, Samples in COLUMNS 
# Remember to set the seed! 
#iNEXT_data <- iNEXT(iNEXT_input_df, 
                   # q = c(0,1,2), datatype = "abundance")

# Save the file
#save(iNEXT_data, file = "data/04_Biodiversity/iNEXT_data.RData")
```

# Evaluate the Diversity! 
```{r load-div}
load("data/04_Biodiversity/iNEXT_data.RData")
str(iNEXT_data)
typeof(iNEXT_data)
```


# Plot Diversity 
```{r plot-rarefaction}
# Prepare Colors 
color_df <- 
  iNEXT_input_df %>%
  colnames() %>%
  data.frame() 
# Check
head(color_df)
# Rename the column 
colnames(color_df)[1] <- "names"
# Check
head(color_df)
color_df <- color_df %>%
  add_row(names = "SRR11364374")

# Make a helper data frame for plotting with colors 
iNEXT_color_df <- 
  color_df %>%
  # Fix the names for merging
  #mutate(names = gsub(names, pattern = "[.]", replace = "-"),
         #names = gsub(names, pattern = "X",  replace = "")) %>%
  # Merge with metadata
  left_join(metadata_df, by = "names") %>%
  # Merge with colors for plotting with ggiNEXT
  left_join(data.frame(treatment_colors = treatment_colors,
                       treatment = names(treatment_colors))) %>%
  mutate(treatment_colors = ifelse(is.na(treatment_colors), "green1", treatment_colors))
            # by = c("names" = "treatment"))  # Corrected variable name for joining

```


# Plot Rarefaction with `ggiNEXT`

```{r ggiNEXT, fig.width=8, fig.height=3.5}
# Plot rarefaction! 
# rarefaction/extrapolation curve, type = 1 

# Order q: 
  # 0 = Richness/ Number of Total taxa
  # 1 = Exponential Shannon / Number of "Common" taxa
  # 2 = Inverse Simpson / Number of "Dominant" taxa 

ggiNEXT(iNEXT_data, type = 1, facet.var = "Order.q") +
 facet_wrap(~Order.q, scales = "fixed") + 
  scale_color_manual(values = iNEXT_color_df$treatment_colors, guide = FALSE) +
  scale_fill_manual(values = iNEXT_color_df$treatment_colors, guide = FALSE) +
  scale_shape_manual(values = rep(1, 60), guide = FALSE) +
  # scale_shape_manual(values = base::rep(17, nsamples(raw_preprocessed_physeq)),
  #                  guide = TRUE) +
  labs(x = "Number of Sequences (Library Size)", 
       y = "Effective Number of ASVs") + 
  theme_bw() + 
  theme(legend.position = "none")
  


```



# Manually plot Diversity 

## Rarefaction
```{r iNEXT-manual, fig.width=6, fig.height=6}
iNEXT_manual_df <- 
  iNEXT_data$iNextEst$size_based %>%
  dplyr::rename(names = Assemblage) %>%
  # Fix the samples names 
  mutate(names = gsub(names, pattern = "[.]", replace = "-"),
         names = gsub(names, pattern = "X", replace = "")) %>%
  filter(names != "SRR11364374") %>%
  # join with metadata 
  left_join(., metadata_df, by = "names") %>%
  # Add colors to data frame
  left_join(., data.frame(treatment_colors = treatment_colors,
                         treatment = names(treatment_colors)),
              by = "treatment") 

unique(iNEXT_manual_df$treatment)
iNEXT_manual_df %>%
  filter(is.na(treatment))

# Inspect 
dim(iNEXT_manual_df)
str(iNEXT_manual_df)

# Plot it - Rarefaction Curve 
iNEXT_manual_df %>%
  # Filter out rows that are calculated by rarefaction from iNEXT
  dplyr::filter(Method == "Rarefaction") %>%
  # Now, let's draw the plot, be sure to group by the samples if using geom_line!
  ggplot(aes(x = m, y= qD, color = treatment, group = names)) + 
  geom_line() + 
  # Facet with the treatment to see the samples better 
  facet_grid(Order.q~treatment, scales = "free") + 
  scale_color_manual(values = treatment_colors) + 
  labs(x = "Number of Sequences (Library Size)", 
       y = "Effective Number of ASVs") + 
  theme_bw() + 
  guides(color = guide_legend(nrow = 2, byrow = TRUE)) + 
  theme(legend.position = "bottom",
        axis.text.x = element_text(angle = 30, vjust = 1, hjust = 1),
        legend.title = element_blank()) 
```

**Interpretation**

Order q of:  

  - 0 = Richness/ Number of Total taxa
  - 1 = Exponential Shannon / Number of "Common" taxa
  - 2 = Inverse Simpson / Number of "Dominant" taxa 
  
From the plots above, For richness, the effective number of ASVs were in the order of  N360 > N240 > N0 > N120 > N180.  Effective number of ASVs changed a little bit cpmaring the treatments from Richness to Shannon to Simpson.

Richness > Shannon > Simpson

Since, this plot was plotted was manually plotted, the axes were different for each of the diversity measure and it seems that the curves plateaued as: Simpson > Shannon > Richness.

Other things that we note from these figures is that, as q increases from 0 to 1 to 2, the abundance of the species is weighted more, so the rarefaction curve plateaus earlier. 
Most of the sample rarefaction curves appear to plateau but we need to look deeper.  

# Boxplots of Diversity 

Now, let's zoom in on a bit of a better view of the data by visualizing it in a boxplot.  

```{r div-dfs}
# Prettier treatment labels 
treatment_names <- c("N0", "N120", "N180", "N240","N360")

names(treatment_names) <- c("N0", "N120", "N180", "N240","N360")

# Make a dataframe
obs_div_df <- 
  iNEXT_manual_df %>%
  dplyr::filter(Method == "Observed") %>%
  left_join(data.frame(treatment_names = treatment_names, 
                      treatment = names(treatment_names)), 
                       by = "treatment")

# Check it 
head(obs_div_df)
```


```{r div-boxplot, fig.height = 3.5, fig.width = 9}
# Boxplots by treatment
obs_div_df %>%
  ggplot(aes(x = treatment_names, y = qD, fill = treatment, color = treatment)) + 
  facet_wrap(~Order.q, scales = "free") + 
  geom_jitter(size = 2.5) + 
  geom_boxplot(alpha = 0.5, outlier.shape = NA) + 
  scale_color_manual(values = treatment_colors) + 
  scale_fill_manual(values = treatment_colors) + 
  labs(y = "Effective Number of ASVs") + 
  theme_bw() +
  theme(legend.position = "bottom",
        axis.title.x = element_blank(),
        legend.title = element_blank())
```

**Interpretation**
For richness, the effective number of ASVs were higher for N120, N0 and N360.  Effective number of ASVs changed a little bit from Richness to Shannon to Simpson. One thing to note is that the scales are also different for all of the three ut in general, the the effective number of ASVs decreased from Richness to Shannon to Simpson. The richness measure of diversity is more sensitive to sequencing depth and compared to Shannon and Simpson indices.



# Diversity vs year of sampling

```{r div-vs-year, fig.height=3.5, fig.width=6}

iNEXT_manual_df %>%
  dplyr::filter(Method == "Observed") %>%
  ggplot(aes(x = year, y = treatment )) + 
  facet_wrap(.~Order.q, scales = "free") + 
  geom_point(aes(color = treatment)) + 
  stat_smooth(method = "lm", formula = y ~poly(x, 2)) + 
  labs(x = "Treatments", y = "# of ASVs") + 
  scale_color_manual(values = treatment_colors) + 
  theme_bw() + 
  theme(legend.position = "bottom",
        legend.title = element_blank()) 
```

Comparing years of sampling with diversity seemed most interesting to me. But quite don not know how to interpret it. Will try different options with this. There were other relationships of the bacterial species with soil properties. Since that data wasnot included in the data set, I can explore these properties when working on my own PhD data.

#Conclusion:

In this file, I learnt how to plot rarefaction curves both manually and with iNEXT. It is interesting to see how different indices look like when plotted. 

# Session Information
```{r session-info}
devtools::session_info()
```
